CMapArea: Resolve sign conversion warning in Draw()
Resolves a trivial sign-conversion warning.

+++ Runtime/AutoMapper/CMapUniverse.cpp
@@ -85,7 +85,7 @@ void CMapUniverse::Draw(const CMapUniverseDrawParms& parms, const zeus::CVector3
 
   int lastWldIdx = -1;
   int lastHexIdx = -1;
-  int instIdx = 0;
+  size_t instIdx = 0;
   for (const CMapObjectSortInfo& info : sortInfos) {
     const CMapWorldData& mwData = x10_worldDatas[info.GetWorldIndex()];
     zeus::CColor surfColor = info.GetSurfaceColor();
